refactor: move LoginData and LoginPayload interfaces to Login.types.ts
