+++ src/shared/types/Login.types.ts
@@ -1,6 +1,27 @@
 import type { AxiosError, AxiosResponse } from "axios";
 import { object, string } from "yup";
 
+export interface LoginData {
+  data: {
+    user: {
+      email: string;
+      firstName: string;
+      lastName: string;
+      middleName: string;
+      _id: string
+      __v: number
+    }
+  }
+  error: null;
+  message: null;
+  success: boolean;
+  version: string;
+}
+
+export interface LoginPayload {
+  email: string;
+  password: string;
+}
 export interface LoginError extends Omit<AxiosError, 'response'> {
   response: AxiosResponse<{
     error: {
